requirements.txt
numpy
argparse
PyVISA-py
